Add Map command to retrieve tilemaps
