+++ UVOCBot/Commands/Utilities/MessageResponseHelpers.cs
@@ -27,10 +27,10 @@ namespace UVOCBot.Commands
             return await _channelAPI.TriggerTypingIndicatorAsync(context.ChannelID, ct).ConfigureAwait(false);
         }
 
-        public async Task<Result<IMessage>> RespondWithEmbedAsync(ICommandContext context, IEmbed embed, CancellationToken ct, Optional<IAllowedMentions> allowedMentions = default)
+        public async Task<Result<IMessage>> RespondWithEmbedAsync(ICommandContext context, IEmbed embed, CancellationToken ct, Optional<FileData> file = default, Optional<IAllowedMentions> allowedMentions = default)
         {
             if (context is InteractionContext ictx)
-                return await RespondToInteractionAsync(ictx, ct, default, new List<IEmbed> { embed }, allowedMentions).ConfigureAwait(false);
+                return await RespondToInteractionAsync(ictx, ct, default, file, new List<IEmbed> { embed }, allowedMentions).ConfigureAwait(false);
             else
                 return await RespondToMessageAsync(context, ct, default, embed: new Optional<IEmbed>(embed), allowedMentions: allowedMentions).ConfigureAwait(false);
         }
@@ -50,7 +50,7 @@ namespace UVOCBot.Commands
                 Colour = BotConstants.DEFAULT_EMBED_COLOUR,
                 Description = content
             };
-            return await RespondWithEmbedAsync(context, embed, ct, allowedMentions).ConfigureAwait(false);
+            return await RespondWithEmbedAsync(context, embed, ct, default, allowedMentions).ConfigureAwait(false);
         }
 
         public async Task<Result<IMessage>> RespondWithUserErrorAsync(ICommandContext context, string content, CancellationToken ct, Optional<IAllowedMentions> allowedMentions = default)
@@ -60,7 +60,7 @@ namespace UVOCBot.Commands
                 Colour = Color.Red,
                 Description = content,
             };
-            return await RespondWithEmbedAsync(context, embed, ct, allowedMentions).ConfigureAwait(false);
+            return await RespondWithEmbedAsync(context, embed, ct, default, allowedMentions).ConfigureAwait(false);
         }
 
         public async Task<Result<IMessage>> RespondWithErrorAsync(ICommandContext context, string content, CancellationToken ct, Optional<IAllowedMentions> allowedMentions = default)
@@ -71,15 +71,16 @@ namespace UVOCBot.Commands
                 Description = content,
                 Footer = new EmbedFooter("Recurring problem? Report it at https://github.com/carlst99/UVOCBot")
             };
-            return await RespondWithEmbedAsync(context, embed, ct, allowedMentions).ConfigureAwait(false);
+            return await RespondWithEmbedAsync(context, embed, ct, default, allowedMentions).ConfigureAwait(false);
         }
 
-        public async Task<Result<IMessage>> RespondToInteractionAsync(InteractionContext context, CancellationToken ct, Optional<string> content = default, Optional<IReadOnlyList<IEmbed>> embeds = default, Optional<IAllowedMentions> allowedMentions = default)
+        public async Task<Result<IMessage>> RespondToInteractionAsync(InteractionContext context, CancellationToken ct, Optional<string> content = default, Optional<FileData> file = default, Optional<IReadOnlyList<IEmbed>> embeds = default, Optional<IAllowedMentions> allowedMentions = default)
         {
             return await _webhookAPI.CreateFollowupMessageAsync(
                 context.ApplicationID,
                 context.Token,
                 content,
+                file: file,
                 embeds: embeds,
                 allowedMentions: allowedMentions,
                 ct: ct).ConfigureAwait(false);
@@ -87,7 +88,16 @@ namespace UVOCBot.Commands
 
         public async Task<Result<IMessage>> RespondToMessageAsync(ICommandContext context, CancellationToken ct, Optional<string> content = default, Optional<string> nonce = default, Optional<bool> isTTS = default, Optional<FileData> file = default, Optional<IEmbed> embed = default, Optional<IAllowedMentions> allowedMentions = default, Optional<IMessageReference> messageReference = default)
         {
-            return await _channelAPI.CreateMessageAsync(context.ChannelID, content, nonce, isTTS, file, embed, allowedMentions, messageReference, ct).ConfigureAwait(false);
+            return await _channelAPI.CreateMessageAsync(
+                context.ChannelID,
+                content,
+                nonce,
+                isTTS,
+                file,
+                embed,
+                allowedMentions,
+                messageReference,
+                ct).ConfigureAwait(false);
         }
     }
 }
