+++ UVOCBot/Model/Census/World.cs
@@ -1,0 +1,15 @@
+﻿namespace UVOCBot.Model.Census
+{
+    public record World
+    {
+        public WorldType WorldId { get; init; }
+        public string State { get; init; }
+        public TranslationProperty Name { get; init; }
+
+        public World()
+        {
+            State = string.Empty;
+            Name = new TranslationProperty();
+        }
+    }
+}
