+++ UVOCBot/Model/Census/WorldType.cs
@@ -1,0 +1,12 @@
+﻿namespace UVOCBot.Model.Census
+{
+    public enum WorldType
+    {
+        Connery = 1,
+        Miller = 10,
+        Cobalt = 13,
+        Emerald = 17,
+        Jaeger = 19,
+        Soltech = 40
+    }
+}
