+++ UVOCBot/Model/Census/ZoneType.cs
@@ -1,0 +1,10 @@
+﻿namespace UVOCBot.Model.Census
+{
+    public enum ZoneType
+    {
+        Indar = 2,
+        Hossin = 4,
+        Amerish = 6,
+        Esamir = 8
+    }
+}
