--- UVOCBot/Model/Planetside/World.cs
@@ -1,15 +1,0 @@
-﻿namespace UVOCBot.Model.Planetside
-{
-    public record World
-    {
-        public WorldType WorldId { get; init; }
-        public string State { get; init; }
-        public TranslationProperty Name { get; init; }
-
-        public World()
-        {
-            State = string.Empty;
-            Name = new TranslationProperty();
-        }
-    }
-}
