--- UVOCBot/Model/Planetside/WorldType.cs
@@ -1,12 +1,0 @@
-﻿namespace UVOCBot.Model.Planetside
-{
-    public enum WorldType
-    {
-        Connery = 1,
-        Miller = 10,
-        Cobalt = 13,
-        Emerald = 17,
-        Jaeger = 19,
-        Soltech = 40
-    }
-}
