--- UVOCBot/Model/Planetside/ZoneType.cs
@@ -1,10 +1,0 @@
-﻿namespace UVOCBot.Model.Planetside
-{
-    public enum ZoneType
-    {
-        Indar = 2,
-        Hossin = 4,
-        Amerish = 6,
-        Esamir = 8
-    }
-}
