+++ UVOCBot/Model/TileMap.cs
@@ -1,0 +1,15 @@
+﻿namespace UVOCBot.Model
+{
+    public enum TileMap
+    {
+        Amerish,
+        Desolation,
+        Esamir,
+        Hossin,
+        Indar,
+        Koltyr,
+        Sanctuary,
+        Tutorial,
+        VR
+    }
+}
