+++ UVOCBot/Services/Abstractions/ICensusApiService.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using UVOCBot.Model.Planetside;
+using UVOCBot.Model.Census;
 
 namespace UVOCBot.Services.Abstractions
 {
